requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.41
asyncpg>=0.30.0
alembic>=1.16.0
pydantic>=2.11.0
pydantic-settings>=2.10.0
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
redis>=6.0.0
celery>=5.5.0
httpx>=0.28.0
email-validator>=2.2.0